+++ Changelog.md
@@ -1,3 +1,17 @@
+**V1.11.13 - Updates**
+- Added end switch support for RA and DEC.
+- Guide pulses are only executed when tracking.
+
+**V1.11.12 - Updates**
+- Change DEC limit code to return degrees instead of stepper positions. 
+- Support separate limits for DEC Limits being queried.
+
+**V1.11.11 - Updates**
+- Fix DEC limit code to use parameters.
+
+**V1.11.10 - Updates**
+- Fix dew heater pin assignments.
+
 **V1.11.9 - Updates**
 - Change storage of RA and DEC steps/degree to be independent of microstepping settings.